+++ src/core/qgsstringstatisticalsummary.h
@@ -159,7 +159,7 @@ class CORE_EXPORT QgsStringStatisticalSummary
      * Returns the number of distinct string values.
      * \see distinctValues()
      */
-    int countDistinct() const { return mValues.keys().count(); }
+    int countDistinct() const { return static_cast<int>( mValues.keys().count() ); }
 
     /**
      * Returns the set of distinct string values.
